+++ pyflp/__init__.py
@@ -160,7 +160,7 @@ def parse(file: pathlib.Path | str) -> Project:
                 if id == PluginID.InternalName:
                     plug_name = event_type(id, value).value
             elif id == PluginID.Data and plug_name is not None:
-                event_type = get_event_by_internal_name(plug_name) or UnknownDataEvent
+                event_type = get_event_by_internal_name(plug_name)
             else:
                 event_type = UnknownDataEvent