+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PLItemBase.offsets` and its fields in `PlaylistEvent` are [float32](https://stackoverflow.com/a/74247360/)
   Thanks to `chrslg` from Stackoverflow and @jubabrut.
 - `Track.height` returns an `str` of its percentage e.g. `100%`.
+- `Instrument.plugin` and `Slot.plugin` return `_PluginBase` for unimplemented
+  native plugins [#102].
 
 ### Fixed
 
@@ -34,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#92]: https://github.com/demberto/PyFLP/issues/92
 [#99]: https://github.com/demberto/PyFLP/issues/99
 [#100]: https://github.com/demberto/PyFLP/issues/100
+[#101]: https://github.com/demberto/PyFLP/issues/101
+[#102]: https://github.com/demberto/PyFLP/issues/102
 
 ## [2.0.0a5.post] - 2022-10-31